Add class methods for components to define options
Component base classes get extended with ComponentOptions which gives
access to a limited version of Thor's DSL for defining command line
options. This is not yet wired in to the CLI, but it automatically
generates a component specific config class that can be used to "hard
code" defaults in a bootstrap definition file.